Partially revert embedding-hub move
