+++ frontend/src/metabase/admin/embedding/embedding-hub/index.ts
@@ -1,3 +1,1 @@
-export * from "./components";
-export * from "./hooks";
-export * from "./types";
+export { EmbeddingHubAdminSettingsPage } from "./EmbeddingHubAdminSettingsPage";
