+++ frontend/src/metabase/embedding/embedding-hub/components/DocsLink.tsx
@@ -18,6 +18,7 @@ export const DocsLink = ({
   utm,
   children,
 }: DocsLinkProps) => {
+  // eslint-disable-next-line metabase/no-unconditional-metabase-links-render -- we are in admin so we must always show the link.
   const { url } = useDocsUrl(docsPath, {
     anchor,
     utm: {
