+++ frontend/src/metabase/embedding/embedding-hub/components/SetupPermissionsAndTenantsPage/DatabaseRoutingStepContent.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable metabase/no-unconditional-metabase-links-render */
+
 import { t } from "ttag";
 
 import { ExternalLink } from "metabase/common/components/ExternalLink";
