+++ frontend/src/metabase/embedding/embedding-hub/components/SetupPermissionsAndTenantsPage/index.ts
