+++ frontend/src/metabase/embedding/embedding-hub/components/SetupPermissionsAndTenantsPage/utils.ts
