+++ frontend/src/metabase/embedding/embedding-hub/components/StepperWithCards/StepperWithCards.tsx
