+++ frontend/src/metabase/embedding/embedding-hub/index.ts
@@ -1,0 +1,3 @@
+export * from "./components";
+export * from "./hooks";
+export * from "./types";
