+++ frontend/src/metabase/embedding/embedding-hub/types/embedding-checklist.ts
