+++ frontend/src/metabase/home/components/EmbeddingHubHomePage/index.ts
@@ -1,0 +1,1 @@
+export { EmbeddingHubHomePage } from "./EmbeddingHubHomePage";
